Добавляет в приложение презентеры
